Feature: Load from stored file

+++ tests/unit/test_key_value_store.py
@@ -8,11 +8,12 @@ class TestKeyValueStore:
 
     @classmethod
     def setup_class(cls):
-        cls.file_path = "db.test"
+        cls.file_path = "test_db.bin"
 
     @classmethod
-    def teardown_class(cls):
-        os.remove(cls.file_path)
+    def teardown_method(cls):
+        if os.path.exists(cls.file_path):
+            os.remove(cls.file_path)
 
     def test_should_read_and_write_in_key_value_store(self):
         kv_store = KeyValueStore(file_path=self.file_path)
@@ -42,5 +43,14 @@ class TestKeyValueStore:
 
         assert kv_store.get_all() == {
             "key_1": {'1': 2, '3': 55, '4': {'1': 2, '3': 55}},
+            "key_3": {'1': 'ss', '3': 55}
+        }
+
+    def test_should_load_existing_file(self):
+        kv_store = KeyValueStore(file_path='tests/unit/test_data_fixture.bin')
+
+        assert kv_store.get_all() == {
+            "key_1": {'1': 2, '3': 55, '4': {'1': 2, '3': 55}},
+            "key_2": {'1': 'amuthan', '3': 55},
             "key_3": {'1': 'ss', '3': 55}
         }
